+++ routes/web.php
@@ -8,20 +8,43 @@ Route::get('/', function () {
     return view('home');
 })->name('home');
 
-Route::get('/contact', function() {
+Route::get('/contact', function () {
     return view('contact');
 })->name('contact');
 
-Route::get('/about', function() {
+Route::get('/about', function () {
     $company = 'Total Automation Solutions';
     return view('about', compact('company'));
 })->name('about');
 
-Route::resource('blueprints', BlueprintController::class);
+// blueprints
+//Route::resource('blueprints', BlueprintController::class);
+Route::get('/blueprints', [BlueprintController::class, 'index'])->name('blueprints.index');
+Route::get('/blueprints/create', [BlueprintController::class, 'create'])
+    ->middleware('auth')
+    ->name('blueprints.create');
+Route::post('/blueprints', [BlueprintController::class, 'store'])
+    ->middleware('auth')
+    ->name('blueprints.store');
+Route::get('/blueprints/{blueprint}', [BlueprintController::class, 'show'])
+    ->name('blueprints.show');
+Route::get('/blueprints/{blueprint}/edit', [BlueprintController::class, 'edit'])
+    ->middleware('auth')
+    ->can('edit', 'blueprint')
+    ->name('blueprints.edit');
+Route::put('/blueprints/{blueprint}', [BlueprintController::class, 'update'])
+    ->middleware('auth')
+    ->can('edit', 'blueprint')
+    ->name('blueprints.update');
+Route::delete('/blueprints/{blueprint}', [BlueprintController::class, 'destroy'])
+    ->middleware('auth')
+    ->can('delete', 'blueprint')
+    ->name('blueprints.destroy');
 
 // blaze
 Route::get('/dashboard', function () {
-    return view('dashboard');
+    $blueprints = Auth::user()->blueprints()->with('category')->get();
+    return view('dashboard', compact('blueprints'));
 })->middleware(['auth', 'verified'])->name('dashboard');
 
 Route::middleware('auth')->group(function () {
@@ -30,4 +53,4 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-require __DIR__.'/auth.php';
+require __DIR__ . '/auth.php';